backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
python-decouple==3.8
asyncpg==0.31.0
alembic==1.17.2
pyjwt==2.10.1
greenlet==3.3.0
aiogram==3.8.0